one method for Marshaling and exclude test structs to own file
