+++ jsonapi/marshal_test.go
@@ -4,133 +4,12 @@ import (
 	"database/sql"
 	"encoding/json"
 	"errors"
-	"fmt"
-	"strconv"
 
 	"gopkg.in/guregu/null.v2/zero"
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 )
-
-type Magic struct {
-	ID MagicID
-}
-
-type MagicID string
-
-func (m MagicID) String() string {
-	return "This should be visible"
-}
-
-type Comment struct {
-	ID   int `json:"-"`
-	Text string
-}
-
-func (c Comment) GetID() string {
-	return fmt.Sprintf("%d", c.ID)
-}
-
-func (c *Comment) SetID(stringID string) error {
-	id, err := strconv.Atoi(stringID)
-	if err != nil {
-		return err
-	}
-
-	c.ID = id
-
-	return nil
-}
-
-type User struct {
-	ID       int
-	Name     string
-	Password string `json:"-"`
-}
-
-func (u User) GetID() string {
-	return fmt.Sprintf("%d", u.ID)
-}
-
-func (u *User) SetID(stringID string) error {
-	id, err := strconv.Atoi(stringID)
-	if err != nil {
-		return err
-	}
-
-	u.ID = id
-
-	return nil
-}
-
-type SimplePost struct {
-	Title, Text string
-}
-
-type Post struct {
-	ID          int
-	Title       string
-	Comments    []Comment     `json:"-"`
-	CommentsIDs []int         `json:"-"`
-	Author      *User         `json:"-"`
-	AuthorID    sql.NullInt64 `json:"-"`
-}
-
-func (c Post) GetID() string {
-	return fmt.Sprintf("%d", c.ID)
-}
-
-func (c *Post) SetID(stringID string) error {
-	id, err := strconv.Atoi(stringID)
-	if err != nil {
-		return err
-	}
-
-	c.ID = id
-
-	return nil
-}
-
-func (c *Post) SetReferencedIDs(ids []ReferenceID) error {
-	return nil
-}
-
-func (c Post) GetReferencedIDs() []ReferenceID {
-	result := []ReferenceID{}
-
-	if c.Author != nil {
-		authorID := ReferenceID{Type: "users", Name: "author", ID: c.Author.GetID()}
-		result = append(result, authorID)
-	} else if c.AuthorID.Valid {
-		authorID := ReferenceID{Type: "users", Name: "author", ID: fmt.Sprintf("%d", c.AuthorID.Int64)}
-		result = append(result, authorID)
-	}
-
-	if len(c.Comments) > 0 {
-		for _, comment := range c.Comments {
-			result = append(result, ReferenceID{Type: "comments", Name: "comments", ID: comment.GetID()})
-		}
-	} else if len(c.CommentsIDs) > 0 {
-		for _, commentID := range c.CommentsIDs {
-			result = append(result, ReferenceID{Type: "comments", Name: "comments", ID: fmt.Sprintf("%d", commentID)})
-		}
-	}
-
-	return result
-}
-
-func (c Post) GetReferencedStructs() []MarshalIdentifier {
-	result := []MarshalIdentifier{c.Author}
-	for key := range c.Comments {
-		result = append(result, &c.Comments[key])
-	}
-	return result
-}
-
-func (c *Post) SetReferencedStructs(references []UnmarshalIdentifier) error {
-	return nil
-}
 
 var _ = Describe("Marshalling", func() {
 	Context("When marshaling simple objects", func() {
@@ -297,7 +176,7 @@ var _ = Describe("Marshalling", func() {
 
 			posts := []Post{post1, post2}
 
-			i, err := MarshalSlice(posts)
+			i, err := Marshal2(posts)
 			Expect(err).To(BeNil())
 
 			expected := map[string]interface{}{
